+++ src/ccm.h
@@ -62,9 +62,9 @@ DPS_Status Encrypt_CCM(const uint8_t key[AES_128_KEY_LENGTH],
                        uint8_t L,
                        const uint8_t nonce[DPS_CCM_NONCE_SIZE],
                        const uint8_t* plainText,
-                       uint32_t ptLen,
+                       size_t ptLen,
                        const uint8_t* aad,
-                       uint32_t aadLen,
+                       size_t aadLen,
                        DPS_TxBuffer* cipherText);
 
 /**
@@ -92,9 +92,9 @@ DPS_Status Decrypt_CCM(const uint8_t key[AES_128_KEY_LENGTH],
                        uint8_t L,
                        const uint8_t nonce[DPS_CCM_NONCE_SIZE],
                        const uint8_t* cipherText,
-                       uint32_t ctLen,
+                       size_t ctLen,
                        const uint8_t* aad,
-                       uint32_t aadLen,
+                       size_t aadLen,
                        DPS_TxBuffer* plainText);
